requirements.txt
PyAudio==0.2.14
python-dotenv==1.1.1
SpeechRecognition==3.14.3
typing_extensions==4.14.0
pyttsx3==2.98
openai==1.93.0
elevenlabs==2.5.0
PyAutoGUI==0.9.54
pillow==11.2.1
tavily==0.7.8
